+++ packages/components/Tabs/src/TabsItem.tsx
@@ -13,7 +13,7 @@ import { TabsContext } from './Tabs';
 import { tabsItemName, TabsItemType, TabsItemProps, TabsItemSlotProps, TabsItemRenderData, TabsItemState } from './TabsItem.types';
 import {
   useAsPressable,
-  useKeyCallback,
+  // useViewCommandFocus,
   createIconProps,
   useOnPressWithFocus,
 } from '@fluentui-react-native/interactive-hooks';
@@ -22,14 +22,13 @@ export const TabsItem = compose<TabsItemType>({
   displayName: tabsItemName,
 
   usePrepareProps: (userProps: TabsItemProps, useStyling: IUseComposeStyling<TabsItemType>) => {
+    const defaultComponentRef = React.useRef(null);
     const {
       icon,
       headerText = '',
-      onAccessibilityTap = userProps.onClick,
       accessibilityLabel = userProps.headerText,
-      componentRef = React.useRef(null),
+      componentRef = defaultComponentRef,
       testID,
-      onClick,
       itemKey,
       itemCount,
       accessibilityPosInSet,
@@ -61,8 +60,6 @@ export const TabsItem = compose<TabsItemType>({
       onFocus: changeSelection,
     });
 
-    const onKeyUp = useKeyCallback(onClick, ' ', 'Enter');
-
     // set up state
     const state: TabsItemState = {
       info: {
@@ -73,6 +70,8 @@ export const TabsItem = compose<TabsItemType>({
         headerText: !!headerText || itemCount !== undefined,
       },
     };
+
+    // const buttonRef = useViewCommandFocus(componentRef);
 
     /* We use the componentRef of the currently selected tabsItem to maintain the default tabbable
     element in Tabs. Since the componentRef isn't generated until after initial render,
@@ -105,7 +104,7 @@ export const TabsItem = compose<TabsItemType>({
         ...rest,
         ...pressable.props,
         ref: componentRef,
-        onAccessibilityTap: onAccessibilityTap,
+        // onAccessibilityTap: onAccessibilityTap,
         accessibilityRole: Platform.OS !== 'windows' ? 'tab' : null, // Add windows role when RN is at >= 0.64
         accessibilityLabel: accessibilityLabel, // For windows, set to read children when RN is at >= 0.64
         accessibilityState: { disabled: userProps.disabled, selected: info.selectedKey === userProps.itemKey },
@@ -113,7 +112,6 @@ export const TabsItem = compose<TabsItemType>({
         accessibilityPositionInSet: accessibilityPosInSet ?? info.tabsItemKeys.findIndex(x => x == itemKey) + 1,
         accessibilitySetSize: accessibilitySetSize ?? info.tabsItemKeys.length,
         onAccessibilityAction: onAccessibilityAction,
-        onKeyUp: onKeyUp,
       },
       content: { children: headerText + countText, testID: testID },
       icon: Platform.OS !== 'windows' ? createIconProps(icon) : null,