+++ packages/components/Tabs/src/TabsItem.types.ts
@@ -5,7 +5,8 @@ import type { IViewProps } from '@fluentui-react-native/adapters';
 import { IFocusable, IPressableState } from '@fluentui-react-native/interactive-hooks';
 import { ITextProps } from '@fluentui-react-native/text';
 import { IconProps } from '@fluentui-react-native/icon';
-import { IButtonProps, IButtonTokens } from '@fluentui-react-native/button';
+import { FontTokens, IBackgroundColorTokens, IBorderTokens, IForegroundColorTokens } from '../../../utils/tokens/lib';
+import { IPressableProps } from '../../Pressable/lib';
 
 export const tabsItemName = 'TabsItem';
 
@@ -35,7 +36,7 @@ export interface TabsItemState {
   info: TabsItemInfo;
 }
 
-export interface TabsItemProps extends IButtonProps {
+export interface TabsItemProps extends IPressableProps {
   /*
    ** The text string for the option
    */
@@ -68,16 +69,41 @@ export interface TabsItemProps extends IButtonProps {
     */
   accessibilitySetSize?: number;
 
+  /*
+   ** An accessibility label for narrator.
+   */
+  accessibilityLabel?: string;
+
   /**
    * A RefObject to access the IFocusable interface. Use this to access the public methods and properties of the component.
    */
   componentRef?: React.RefObject<IFocusable>;
+
+  /**
+   * Source URL or name of the icon to show on the TabsItem.
+   */
+  icon?: IconSourcesType;
 }
 
-export interface TabsItemTokens extends IButtonTokens {
-  textBorderColor?: string;
+type IconSourcesType = number | string | IconProps;
+
+export interface TabsItemTokens extends IForegroundColorTokens, FontTokens, IBackgroundColorTokens, IBorderTokens {
+  /**
+   * The indicator color.
+   */
   indicatorColor?: string;
+
+  /**
+   * The icon color.
+   */
+  iconColor?: string;
+
+  /**
+   * Source URL or name of the icon to show on the TabsItem.
+   */
+   icon?: IconSourcesType;
 }
+
 
 export interface TabsItemSlotProps {
   root: React.PropsWithRef<IViewProps>;